# apps/brain/api/openapi.yaml
openapi: 3.1.0
info:
  title: Openarity brain
  version: 0.1.0
  description: |
    The brain's HTTP API.

    This document is the contract between the service and everything that
    talks to it, including the CLI. It is hand-written and reviewed as a diff:
    a change here is a change to what callers may rely on.

    A test compares the routes the service registers against the paths below.
    Adding an endpoint without describing it here fails the build.

    **Errors are `text/plain`, not JSON.** Successful responses are JSON;
    failures carry a short sentence and nothing machine-readable. Clients
    should branch on the status code, never on the body.
  license:
    name: MIT
    url: https://github.com/LaplacianAI/openarity/blob/main/LICENSE
servers:
  - url: http://127.0.0.1:21120
    description: The API listener, OPENARITY_API_BIND
...
security:
  - bearerAuth: []
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams
      description: |
        A super admin sees every team, paged. Everyone else sees the teams
        they belong to, with their role in each, in a single page with no
        cursor — a person's own memberships are bounded.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"

    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      description: |
        Super admins only. Creating a team is platform-scoped, so there is no
        team to check a role against.

        The new team has no members. A super admin adds the first one, who is
        then able to manage the rest.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: The team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A team already has that name
          content:
            text/plain:
              schema:
                type: string
                examples: ["a team with that name already exists"]
  /teams/{id}:
    parameters:
      - $ref: "#/components/parameters/TeamID"
    get:
      operationId: getTeam
      tags: [teams]
      summary: Read one team
      description: |
        Members and super admins only. A team the caller may not see and a
        team that does not exist answer identically — a 403 would confirm the
        id is real.
      responses:
        "200":
          description: The team, carrying the caller's role if they are a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /teams/{id}/members:
    parameters:
      - $ref: "#/components/parameters/TeamID"
    get:
      operationId: listTeamMembers
      tags: [teams]
      summary: List a team's members
      description: |
        A read, so seeing the team is enough — no action is required.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a member
      description: |
        Requires `membership:write` in this team, and nothing else — naming
        somebody by `subject` deliberately does not require permission to read
        the directory, because "add the person I can name" is a smaller
        authority than "show me everyone".

        The role is checked by the database rather than the service: roles are
        rows, so an unknown one is a rejected foreign key and comes back as a
        400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "204":
          description: Added
        "400":
          description: |
            The body is malformed, names a user or role that does not exist,
            or carries neither `user_id` nor `subject` — or both.
          content:
            text/plain:
              schema:
                type: string
                examples: ["unknown role"]
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: |
            No such team, or none the caller may see. A `subject` nobody has
            is a 404 too: the caller named a person, and there is no person.
          content:
            text/plain:
              schema:
                type: string
                examples: ["no user has that subject"]
        "409":
          description: |
            Already a member, or the `subject` matched more than one person.

            The second needs no federation, which an earlier version of this
            description got wrong. A user row is keyed on `(issuer, subject)`
            and rows accumulate, so one provider whose issuer URL changed —
            moved behind a tunnel, addressed by IP rather than by name — leaves
            the same person behind under both. Observed in a single-provider
            deployment.

            The reply names the issuer beside each id, because the id alone
            gives no basis to choose and the wrong choice grants a role to an
            identity nobody can authenticate as.
          content:
            text/plain:
              schema:
                type: string
                examples:
                  - "already a member"
                  - "2 users have that subject — retry with user_id:
                     47c93149-6e9e-49ea-a9f6-34011f1e1e04 (https://auth.example.com/),
                     e096f85a-a05b-4699-919a-e0c5c5b57072 (https://auth.internal/)"
  /teams/{id}/members/{userID}:
    parameters:
      - $ref: "#/components/parameters/TeamID"
      - name: userID
        in: path
        required: true
        description: The user to remove
        schema:
          type: string
          format: uuid
    delete:
      operationId: removeTeamMember
      tags: [teams]
      summary: Remove a member
      description: |
        Requires `membership:write` in this team. Removing someone who is not a
        member succeeds: the caller asked for a state, and that state holds.
      responses:
        "204":
          description: No longer a member
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Find a user
      description: |
        Adding someone to a team needs their id, and an id is the one thing
        nobody can tell you over chat. This is where it is looked up.

        Requires `user:read` in at least one team, and super admins always.
        Not every authenticated caller: the directory is every subject and
        email the deployment has ever seen, and a filtered read of it is an
        existence oracle for anyone who can guess a username.

        The check is "somewhere" rather than "in this team", because the
        caller is looking someone up precisely to find out which team they
        should be in — there is no team to check against yet.

        `subject` narrows to an exact match. There is no prefix or substring
        search, deliberately: a directory that answers "who starts with s" is
        one someone walks.
      parameters:
        - name: subject
          in: query
          required: false
          description: |
            The identifier at the provider, matched exactly. An unknown one is
            an empty page, not a 404.
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        An OIDC access token, or the shared development token when
        OPENARITY_DEV_TOKEN is set. The scheme is matched case-insensitively.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: |
        Rows per page. A value above the maximum is clamped rather than
        refused; zero, a negative and anything unparseable are 400.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        An opaque position, taken verbatim from the `next_cursor` of the
        previous page. It is not constructible by a client, and one that has
        been altered is a 400 rather than a silent restart.
      schema:
        type: string
    TeamID:
      name: id
      in: path
      required: true
      description: The team
      schema:
        type: string
        format: uuid
  responses:
    Unauthenticated:
      description: |
        No token, or one that no configured verifier accepts. Carries a
        `WWW-Authenticate: Bearer` header.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            examples: ["Bearer"]
      content:
        text/plain:
          schema:
            type: string
            examples: ["unauthorized"]
    Forbidden:
      description: |
        The token is valid; the caller may not do this. They should not be
        asked to authenticate again.
      content:
        text/plain:
          schema:
            type: string
            examples: ["forbidden"]
    NotFound:
      description: |
        No such resource, or none the caller may see. The two are
        indistinguishable on purpose.
      content:
        text/plain:
          schema:
            type: string
            examples: ["not found"]
    BadRequest:
      description: |
        The request is malformed. The reply names the problem and never echoes
        the value back, because a body can carry a credential.
      content:
        text/plain:
          schema:
            type: string
            examples: ["id must be a uuid"]
  schemas:
...
    Team:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
          examples: ["platform"]
        role:
          type: string
          description: |
            The caller's role in this team. Absent when they are not a member,
            which is the normal case for a super admin.
          examples: ["admin"]
    Member:
      type: object
      required: [user_id, subject, role]
      properties:
        user_id:
          type: string
          format: uuid
        subject:
          type: string
          description: The member's identifier at their provider
        email:
          type: string
          format: email
          description: Absent when the provider released none
        role:
          type: string
          examples: ["member"]
    CreateTeamRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: |
            Trimmed before it is stored, and unique across teams. A name of
            only whitespace is a 400.
    AddMemberRequest:
      type: object
      required: [role]
      additionalProperties: false
      description: |
        Exactly one of `user_id` and `subject`. Neither is a 400, and so is
        both — guessing which the caller meant is how somebody ends up in a
        team they were never named for.

        `subject` exists so that adding a person you can name does not require
        permission to read the whole directory. It is the smaller authority,
        and it is the one this endpoint should need.

        The exactly-one rule is enforced by the service and stated here rather
        than written as a `oneOf`: oapi-codegen turns a top-level `oneOf` into
        a `json.RawMessage` union with a hand-rolled `MarshalJSON`, and every
        client then builds this body through generated accessors instead of a
        struct literal. Verified, not assumed.
      properties:
        user_id:
          type: string
          format: uuid
          description: |
            Unambiguous, and the only form that works when two providers issue
            the same subject. `GET /users` is where one comes from.
        subject:
          type: string
          description: |
            The identifier at the provider, matched exactly. Unique only per
            issuer, and rows survive a change of issuer, so this can match more
            than one person without a second provider ever being configured —
            that answers 409, never a guess.
          examples: ["alice"]
        role:
          type: string
          description: |
            Must name a row in `roles`. Ships with `admin` and `member`;
            deployments may add their own.
          examples: ["member"]
    TeamPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Team"
        next_cursor:
          type: string
          description: |
            Absent on the last page. Its absence is the only end-of-collection
            signal — there is no separate flag to disagree with it.
    MemberPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        next_cursor:
          type: string
          description: Absent on the last page
...
    User:
      type: object
      required: [id, issuer, subject]
      properties:
        id:
          type: string
          format: uuid
        issuer:
          type: string
          description: |
            The identity provider that issued their token, from the `iss`
            claim. Identity is `(issuer, subject)`, so this is not decoration:
            it is the half that makes two rows with the same subject two
            different people. Without it a listing cannot be read.
          examples: ["https://auth.example.com/application/o/openarity/"]
        subject:
          type: string
          description: The identifier at their provider
          examples: ["akadmin"]
        email:
          type: string
          format: email
          description: Absent when the provider released none
    UserPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        next_cursor:
          type: string
          description: Absent on the last page