# apps/brain/internal/api/users/users.go
package users

import (
	"context"
	"log/slog"
	"net/http"
	"strings"

	"github.com/LaplacianAI/openarity/apps/brain/internal/api"
	"github.com/LaplacianAI/openarity/apps/brain/internal/store/db"
)

type Store interface {
	ListUsers(ctx context.Context, arg db.ListUsersParams) ([]db.ListUsersRow, error)
}

type handler struct {
	logger *slog.Logger
	store  Store
}

func New(logger *slog.Logger, s Store) *api.Router {
	h := &handler{logger: logger, store: s}

	r := api.NewRouter("/users")
	r.Get("", h.list)
	return r
}

func (h *handler) list(w http.ResponseWriter, r *http.Request) {
	u := api.Caller(r)

	limit, ok := api.Limit(w, r)
	if !ok {
		return
	}

	params, ok := h.userPage(w, r, limit)
	if !ok {
		return
	}

	rows, err := h.store.ListUsers(r.Context(), params)
	if err != nil {
		api.Fail(w, h.logger, u, "failed to list users", err)
		return
	}

	page, err := api.MapPage(rows, limit,
		func(row db.ListUsersRow) any { return userCursor{Subject: row.Subject, ID: row.ID} },
		func(row db.ListUsersRow) user {
			return user{
				ID:      row.ID,
				Issuer:  row.Issuer,
				Subject: row.Subject,
				Email:   row.Email,
			}
		},
	)
	if err != nil {
		api.Fail(w, h.logger, u, "failed to page users", err)
		return
	}

	api.WriteJSON(w, h.logger, http.StatusOK, page)
}

func (h *handler) userPage(w http.ResponseWriter, r *http.Request, limit int32) (db.ListUsersParams, bool) {
	params := db.ListUsersParams{PageSize: limit + 1}

	if subject := strings.TrimSpace(r.URL.Query().Get("subject")); subject != "" {
		params.UseSubject = true
		params.Subject = subject
	}

	raw := r.URL.Query().Get("cursor")
	if raw == "" {
		return params, true
	}

	var c userCursor
	if !api.DecodeCursor(w, raw, &c) {
		return params, false
	}

	params.UseCursor = true
	params.AfterSubject = c.Subject
	params.AfterID = c.ID
	return params, true
}